Split game application into modules
